test(cli): merge Atlassian API token unit tests into two modules
