--- app/src/context-engine/tests/unit/test_credentials_store_integrations.py
@@ -1,157 +1,0 @@
-"""Integration credential paths in credentials_store."""
-
-from __future__ import annotations
-
-import pytest
-
-from adapters.inbound.cli import credentials_store as cs
-
-
-@pytest.fixture
-def keychain(monkeypatch: pytest.MonkeyPatch) -> dict[tuple[str, str], str]:
-    stored: dict[tuple[str, str], str] = {}
-
-    def set_password(service: str, username: str, password: str) -> None:
-        stored[(service, username)] = password
-
-    def get_password(service: str, username: str) -> str | None:
-        return stored.get((service, username))
-
-    def delete_password(service: str, username: str) -> None:
-        stored.pop((service, username), None)
-
-    monkeypatch.setattr(cs.keyring, "set_password", set_password)
-    monkeypatch.setattr(cs.keyring, "get_password", get_password)
-    monkeypatch.setattr(cs.keyring, "delete_password", delete_password)
-    return stored
-
-
-def test_linear_integration_tokens_roundtrip(
-    monkeypatch: pytest.MonkeyPatch, tmp_path, keychain: dict
-) -> None:
-    monkeypatch.setattr(cs, "credentials_path", lambda: tmp_path / "credentials.json")
-    cs.save_integration_tokens(
-        "linear",
-        {
-            "access_token": "access",
-            "refresh_token": "refresh",
-            "scope": "read",
-            "expires_at": 9999999999.0,
-        },
-    )
-    tokens = cs.get_integration_tokens("linear")
-    assert tokens["access_token"] == "access"
-    assert tokens["refresh_token"] == "refresh"
-
-    cs.clear_integration_tokens("linear")
-    assert cs.get_integration_tokens("linear") == {}
-
-
-def test_list_integration_providers(
-    monkeypatch: pytest.MonkeyPatch, tmp_path, keychain: dict
-) -> None:
-    monkeypatch.setattr(cs, "credentials_path", lambda: tmp_path / "credentials.json")
-    cs.save_integration_tokens(
-        "linear",
-        {"access_token": "tok", "expires_at": 9999999999.0},
-    )
-    providers = cs.list_integration_providers()
-    assert "linear" in providers
-
-
-def test_clear_integration_tokens_linear(
-    monkeypatch: pytest.MonkeyPatch, tmp_path, keychain: dict
-) -> None:
-    monkeypatch.setattr(cs, "credentials_path", lambda: tmp_path / "credentials.json")
-    cs.save_integration_tokens(
-        "linear",
-        {"access_token": "a", "refresh_token": "r", "expires_at": 9999999999.0},
-    )
-    cs.clear_integration_tokens("linear")
-    assert cs.get_integration_tokens("linear") == {}
-    status = cs.get_integration_status("linear")
-    assert status["authenticated"] is False
-
-
-def test_readable_credentials_path_prefers_legacy_file(
-    monkeypatch: pytest.MonkeyPatch, tmp_path
-) -> None:
-    cfg = tmp_path / "xdg"
-    legacy_cfg = tmp_path / "legacy-xdg"
-    monkeypatch.setenv("XDG_CONFIG_HOME", str(cfg))
-    monkeypatch.setattr(cs, "legacy_config_dir", lambda: legacy_cfg)
-    legacy_cfg.mkdir(parents=True)
-    legacy_file = legacy_cfg / "credentials.json"
-    legacy_file.write_text('{"api_key": "from-legacy"}', encoding="utf-8")
-
-    assert cs.readable_credentials_path() == legacy_file
-    assert cs.read_credentials().get("api_key") == "from-legacy"
-
-
-def test_read_credentials_invalid_json_returns_empty(
-    monkeypatch: pytest.MonkeyPatch, tmp_path
-) -> None:
-    monkeypatch.setattr(cs, "credentials_path", lambda: tmp_path / "credentials.json")
-    path = tmp_path / "credentials.json"
-    path.write_text("{not json", encoding="utf-8")
-    assert cs.read_credentials() == {}
-
-
-def test_get_integration_status_unknown_provider(
-    monkeypatch: pytest.MonkeyPatch, tmp_path
-) -> None:
-    monkeypatch.setattr(cs, "credentials_path", lambda: tmp_path / "credentials.json")
-    with pytest.raises(ValueError, match="Unknown integration provider"):
-        cs.get_integration_status("github")
-
-
-def test_linear_status_includes_org_and_scope_string(
-    monkeypatch: pytest.MonkeyPatch, tmp_path, keychain: dict
-) -> None:
-    monkeypatch.setattr(
-        "adapters.inbound.cli.integration_profile.fetch_linear_viewer",
-        lambda _token: {
-            "account": {"name": "Ada", "email": "ada@example.com"},
-            "organization": {"name": "Potpie"},
-        },
-    )
-    monkeypatch.setattr(cs, "credentials_path", lambda: tmp_path / "credentials.json")
-    cs.save_integration_tokens(
-        "linear",
-        {
-            "access_token": "access",
-            "refresh_token": "refresh",
-            "expires_at": 9999999999.0,
-            "scope": "read,write",
-            "account": {"name": "Ada", "email": "ada@example.com"},
-            "organization": {"name": "Potpie"},
-        },
-    )
-    status = cs.get_integration_status("linear")
-    assert status["authenticated"] is True
-    assert status["site_name"] == "Potpie"
-    assert status["scope"] == ["read", "write"]
-
-
-def test_store_secure_secret_generic_exception(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    def set_password(service: str, username: str, password: str) -> None:
-        raise RuntimeError("unexpected")
-
-    monkeypatch.setattr(cs.keyring, "set_password", set_password)
-    with pytest.raises(cs.CredentialStoreError, match="keychain"):
-        cs.store_secure_secret("name", "secret")
-
-
-def test_load_secure_secret_generic_exception(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    def get_password(service: str, username: str) -> str | None:
-        raise RuntimeError("unexpected")
-
-    monkeypatch.setattr(cs.keyring, "get_password", get_password)
-    with pytest.raises(cs.CredentialStoreError, match="keychain"):
-        cs.load_secure_secret("name")
-
-
